Feat/add to favs

+++ apps/web/app/routes/($lang).logout.jsx
@@ -1,9 +1,17 @@
 import { redirect } from "@remix-run/node";
 import { sessionStorage } from "@pronto/auth/auth.server";
+import authenticator from "@pronto/auth/auth.server";
 
 export const loader = async ({ request, params }) => {
   const locale = params.lang || "en";
-  const session = await sessionStorage.getSession(request.headers.get("Cookie"));
+  const user = await authenticator.isAuthenticated(request);
+
+  if (!user) {
+    return redirect(`/${locale}/login`);
+  }
+  const session = await sessionStorage.getSession(
+    request.headers.get("Cookie"),
+  );
   return redirect(`/${locale}/login`, {
     headers: {
       "Set-Cookie": await sessionStorage.destroySession(session),
